requirements.txt
django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-unfold==0.32.0